requirements.txt
pyyaml
matplotlib
networkx
opencv-python
Pillow
numpy
ipywidgets
tqdm
Cython
pandas
sortedcontainers
file-read-backwards
